refactor: rewrite sample page
