+++ app/Http/Controllers/SampleController.php
@@ -4,13 +4,13 @@ namespace App\Http\Controllers;
 
 use Inertia\Inertia;
 
-class TestController extends Controller
+class SampleController extends Controller
 {
     /**
      * @return \Inertia\Response
      */
     public function __invoke()
     {
-        return Inertia::render('test');
+        return Inertia::render('sample/index');
     }
 }
