+++ resources/lang/en/sample.php
@@ -1,0 +1,9 @@
+<?php
+
+return [
+    'menu' => 'Sample Menu',
+
+    'routes' => [
+        'index' => 'Sample Page',
+    ],
+];
