+++ resources/lang/id/sample.php
@@ -1,0 +1,9 @@
+<?php
+
+return [
+    'menu' => 'Contoh Menu',
+
+    'routes' => [
+        'index' => 'Contoh Halaman',
+    ],
+];
